+++ app/windows/runner/flutter_window.h
@@ -2,7 +2,9 @@
 #define RUNNER_FLUTTER_WINDOW_H_
 
 #include <flutter/dart_project.h>
+#include <flutter/encodable_value.h>
 #include <flutter/flutter_view_controller.h>
+#include <flutter/method_channel.h>
 
 #include <memory>
 
@@ -28,6 +30,10 @@ class FlutterWindow : public Win32Window {
 
   // The Flutter instance hosted by this window.
   std::unique_ptr<flutter::FlutterViewController> flutter_controller_;
+
+  // Channel used to forward OS file drops (WM_DROPFILES) to Dart.
+  std::unique_ptr<flutter::MethodChannel<flutter::EncodableValue>>
+      drop_channel_;
 };
 
 #endif  // RUNNER_FLUTTER_WINDOW_H_